Solved bug when copying SceneParts, used for save/load .objs from cache.
When copying SceneParts, a new set of smartpointers were created in every primitive to the ScenePart it belongs. So, later, every time a smartpointer pointing to ScenePart was assigned to a primitive in XmlAssetsLoader:798, each assignation destroys smartpointer and assigns the new one pointing to the same address. In the last iteration, when the last smartpointer is destroyed to make room to the copy, the ScenePart itself is destroyed and everything blows up.