Hashing objs (#53)
* Initial implementation of OBJ models cache.

* Solved bug when copying SceneParts, used for save/load .objs from cache.

When copying SceneParts, a new set of smartpointers were created in every primitive to the ScenePart it belongs. So, later, every time a smartpointer pointing to ScenePart was assigned to a primitive in XmlAssetsLoader:798, each assignation destroys smartpointer and assigns the new one pointing to the same address. In the last iteration, when the last smartpointer is destroyed to make room to the copy, the ScenePart itself is destroyed and everything blows up.

* Added WavefrontObjCache destructor.

* Splitted loadObj() into several smaller functions.

* Added initial implementation of a class to represent a loaded .obj file.

* Adapted WavefrontObjCache to work with WavefrontObj.

* Improved the way the loaded .obj is saved in the cache. Fixed obscure memory leak when cloning triangles.

* Corrected primitives count.

* Improved cache system. Now, only a maximum number of elements are allowed to be in the cache (32 by defaut)

* Changed verbosity of obj filter related messages. Added scenepart counter.

* Added time measurements by means of TimeWatcher class. Changed verbosity of several messages.

Co-authored-by: Lukas Winiwarter <lukas.winiwarter@uni-heidelberg.de>